refactor(canvas): split chat constants by owner
