+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatMentionPopup/index.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef } from 'react';
 import './index.css';
-import { MENTION_GROUP_LABEL_KEY, getMentionGroupKey } from '../constants';
+import { MENTION_GROUP_LABEL_KEY, getMentionGroupKey } from './constants';
 import type { MentionItem } from '../types';
 import { MentionNodeIcon, tabMentionIconType } from '../utils/mentions';
 import { roleColorSoft } from '../../../utils/roleColors';
