+++ apps/canvas-workspace/src/renderer/src/components/chat/SessionTitle.tsx
@@ -1,4 +1,4 @@
-import { DOM_MENTION_PREFIX } from './constants';
+import { DOM_MENTION_PREFIX } from './ChatMentionPopup/constants';
 import { MentionNodeIcon } from './utils/mentions';
 import { sessionTitleParts, sessionTitleText } from './utils/sessionTitle';
